requirements.txt
selenium
kiteconnect
pyrebase4
schedule